+++ src/ui/VirtualRenderers/ColorBar.test.tsx
