+++ src/ui/VirtualRenderers/VirtualDiffRenderer.tsx
